+++ src/components/card/inventory/ItemName.jsx
@@ -1,5 +1,5 @@
 import { Card } from "@radix-ui/themes";
 
 export const ItemNameCard = ({ itemObject }) => {
-    return <Card>{itemObject.name}</Card>;
+    return <Card m="2">{itemObject.name}</Card>;
 };
